修改学生信息查询接口

+++ ggs-dal/src/main/java/io/chicken/ggs/dal/dao/BaseDao.java
@@ -26,7 +26,7 @@ public interface BaseDao<T> {
 
     List<T> queryList(Object id);
 
-    int queryTotal(Map<String, Object> map);
+    long queryTotal(Map<String, Object> map);
 
     int queryTotal();
 }
